方法放到models.py中
